requirements.txt
Django==1.10.2
django-cors-middleware==1.3.1
django-filter==0.15.2
django-guardian==1.4.6
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
Pillow==3.4.1
psycopg2==2.6.2
PyJWT==1.4.2
six==1.10.0
